UpdateResolution.sh: - 4:3 aspect ratio command implementation - 16:9 aspect ratio command implementation - unified refresh[+-] and resolution[+-] commands and enhanced
AVWizardWrapper.sh:
- updated to use new StartX function
- added section for step 3 (resolution) wizard interrupt

AVWizard-Common.sh:
- added "Resolution_GetFullName" function
- changed StartX and StopX functions to use the new Start/Stop_X.sh scripts to use AVWizard as the xinit client, w/o a window manager
- removed SDL dummy driver variable

RunWizard.sh:
- removed X starting code

+++ src/AVWizard/scripts/UpdateResolution.sh
@@ -5,36 +5,79 @@
 Param="$1"
 case "$Param" in
 	4[_:]3)
-		echo "--> Aspect ratio 4:3"
-		exit 0; # Already in 4:3; no change
+		Video_Ratio=$(WizGet Video_Ratio)
+		echo "--> Aspect ratio 4:3 (was $Video_Ratio)"
+		if [[ "$Video_Ratio" == 4_3 ]]; then
+			exit 0 # No change required
+		fi
+		kill -USR1 $PPID
+		"$BaseDir"/Xconfigure.sh --conffile "$XF86Config" --defaults --resolution '640x480'
+		WizSet Video_Ratio '4_3'
+		WizSet VideoResolution '640x480'
+		WizSet VideoRefresh '60'
+		WizSet WindowWidth 640
+		WizSet WindowHeight 480
 	;;
 	16[_:]9)
-		echo "--> Aspect ratio 16:9"
+		Video_Ratio=$(WizGet Video_Ratio)
+		echo "--> Aspect ratio 16:9 (was $Video_Ratio)"
+		if [[ "$Video_Ratio" == 16_9 ]]; then
+			exit 0 # No change required
+		fi
+		kill -USR1 $PPID
+		Video_Ratio=$(WizGet Video_Ratio)
+		"$BaseDir"/Xconfigure.sh --conffile "$XF86Config" --defaults --resolution '848x480'
 		WizSet Video_Ratio '16_9'
 		WizSet VideoResolution '480p'
 		WizSet VideoRefresh '60'
-		"$BaseDir"/Xconfigure.sh --conffile "$XF86Config" --defaults --resolution '848x480'
-		RestartX
-		kill -USR1 $PPID
+		WizSet WindowWidth 848
+		WizSet WindowHeight 480
 	;;
-	resolution[+-])
+	resolution[+-]|refresh[+-])
 		Video_Ratio=$(WizGet Video_Ratio)
 		VideoResolution=$(WizGet VideoResolution)
-		echo "--> Resolution cmd: $Param; Ratio: $Video_Ratio; Current resolution: $VideoResolution"
+		VideoRefresh=$(WizGet VideoRefresh)
+		echo "--> Cmd: $Param; Ratio: $Video_Ratio; Current resolution: $VideoResolution@$VideoRefresh"
 		case "${Param#resolution}" in
 			+) VideoResolution_New=$(Resolution_Next "$Video_Ratio" "$VideoResolution") ;;
 			-) VideoResolution_New=$(Resolution_Prev "$Video_Ratio" "$VideoResolution") ;;
 		esac
-		echo "--> New resolution: $VideoResolution_New"
-	;;
-	refresh[+-])
-		VideoRefresh=$(WizGet VideoRefresh)
-		echo "--> Refresh cmd: $Param; Ratio: $Video_Ratio; Current refresh: $VideoRefresh"
 		case "${Param#refresh}" in
 			+) VideoRefresh_New=$(Refresh_Next "$VideoRefresh") ;;
 			-) VideoRefresh_New=$(Refresh_Prev "$VideoRefresh") ;;
 		esac
-		echo "--> New refresh: $VideoRefresh_New"
+
+		if [[ -z "$VideoResolution_New" ]]; then
+			VideoResolution_New="$(Resolution_GetFullName "$Video_Ratio" "$VideoResolution")"
+		fi
+		if [[ -z "$VideoRefresh_New" ]]; then
+			VideoRefresh_New="$VideoRefresh"
+		fi
+
+		echo "--> New resolution: $VideoResolution_New (was: $VideoResolution)"
+		echo "--> New refresh: $VideoRefresh_New (was: $VideoRefresh)"
+
+		if [[ "$VideoResolution_New" == *=* ]]; then
+			VideoResolution_Name="${VideoResolution_New%%=*}"
+			VideoResolution_Size="${VideoResolution_New#*=}"
+		else
+			VideoResolution_Name="$VideoResolution_New"
+			VideoResolution_Size="$VideoResolution_New"
+		fi
+
+		echo [[ "$VideoResolution" "$VideoResolution_Name" "$VideoRefresh" "$VideoRefresh_New" ]]
+		if [[ "$VideoResolution" == "$VideoResolution_Name" && "$VideoRefresh" == "$VideoRefresh_New" ]]; then
+			exit 0 # No change required
+		fi
+		kill -USR1 $PPID
+
+		"$BaseDir"/Xconfigure.sh --conffile "$XF86Config" --defaults --resolution "$VideoResolution_Size@$VideoRefresh_New"
+		WizSet VideoResolution "$VideoResolution_Name"
+		WizSet VideoRefresh "$VideoRefresh_New"
+		WindowWidth="${VideoResolution_Size%x*}"
+		WindowHeight="${VideoResolution_Size#*x}"
+		WizSet WindowWidth "$WindowWidth"
+		WizSet WindowHeight "$WindowHeight"
 	;;
 	*) echo "Unknown parameter: '$Param'"; exit 1 ;;
 esac
